Increased replica factor
Looks like there is a problem with swarm overlay network resolving again the IP of a container after a rolling update of a service. By providing more than one replica in a cluster of 2 we will ensure there's always going to exist a container with the service on each host. This should be investigated.